added FAST I2c
Fast I2c allowed to stop errors from happening, this could be a sign
that the 100 ms time for the entire run-loop is just too fast for the
system in general. If so in particular local variables that are
recreated often might cause issues. Switching these to global might be
a solution